+++ tests/common/test_constants.py
@@ -1,0 +1,7 @@
+"""Test constants."""
+
+# Punch Clock Imports
+from punchclock.common.constants import getConstants
+
+constants = getConstants()
+print(f"{constants=}")